astro/vercel
Forced install of @astrojs/vercel. This may be a breaking change, as the npx install suggests that version 5 of astro was required. As we couldn't install version 5 beta without issues with Tailwind, opted to leave it at version 4. This SHOULD still work. I've used the vercel adapter before in other Astro 4 projects, so I assume this is just a bug that hasn't been addressed.